chatting/backend/requirements.txt
appnope==0.1.3
asgiref==3.5.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
backcall==0.2.0
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
constantly==15.1.0
cryptography==38.0.4
daphne==4.0.0
decorator==5.1.1
Django==4.1.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
executing==1.2.0
Faker==15.3.4
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipython==8.7.0
jedi==0.18.2
matplotlib-inline==0.1.6
msgpack==1.0.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
pyOpenSSL==22.1.0
python-dateutil==2.8.2
python-decouple==3.7
pytz==2022.6
redis==4.4.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.6.0
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
wcwidth==0.2.5
zope.interface==5.5.2
